fixed tests for JUnit result parser
